requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-CORS==4.0.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1
requests==2.31.0
ShopifyAPI==12.3.0
WooCommerce==3.0.0
python-dotenv==1.0.0
openai==0.28.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
plotly==5.15.0
gunicorn==21.2.0
pytest==7.4.0
pytest-flask==1.2.0
